+++ src/main.rs
@@ -488,6 +488,10 @@ fn dispatch_common_subcommand(
             let result = commands::delete_vhost(client, second_level_args);
             res_handler.delete_operation_result(result);
         }
+        ("nodes", "list") => {
+            let result = commands::list_nodes(client);
+            res_handler.tabular_result(result)
+        }
         ("purge", "queue") => {
             let result = commands::purge_queue(client, &vhost, second_level_args);
             res_handler.no_output_on_success(result);